Clean commit - train, test, trading, ranking working
